+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Uliweb Change Log
   it'll limit only class object can invoke the method rather than instance object.
 * Refactor upload app, add `download` to functions
 * Improve secretkey app, add `-o` to specify output file. Add `keyfile` to most functions.
+* Add `MIME_TYPES` section to upload app, but it'll only effect with uliweb application, but not web server
 
 0.1.7 Version
 -----------------